requirements.txt
boto==2.38.0
bz2file==0.98
cairocffi==0.6
cffi==0.9.2
decorator==3.4.0
gensim==0.11.1.post1
httplib2==0.9
matplotlib==1.4.3
mock==1.0.1
networkx==1.9.1
nltk==3.0.2
nose==1.3.4
numpy==1.9.2
pandas==0.16.0
pycparser==2.10
pygraphviz==1.2
pymongo==3.0
pyparsing==2.0.3
python-dateutil==2.4.1
python-instagram==1.3.0
pytz==2014.10
scipy==0.15.1
seaborn==0.5.1
simplejson==3.6.5
six==1.9.0
smart-open==1.2.1
